refactor: refactor unit tests
