+++ src/constants.ts
@@ -19,7 +19,12 @@ export const DEFAULT_SGID_CODE_CHALLENGE_METHOD = 'S256'
 export const API_VERSION = 2
 
 /**
+ * Testing
+ */
+/**
  * Regex pattern that the code verifier and code challenge in the PKCE flow should match according to the PKCE RFC
  * https://www.rfc-editor.org/rfc/rfc7636
  */
 export const codeVerifierAndChallengePattern = /^[A-Za-z\d\-._~]{43,128}$/
+
+export const DEFAULT_RESPONSE_TYPE = 'code'
